A first, rudimentary implementation of template arguments. examples/chatrooms.fun is a good example of template argument usage that is possible now.
A template argument can receive its value at runtime (e.g. if calling a template from inside of a for loop and passing in the iterator value as an argument). The runtime value of a template argument gets calculated at template invocation time; the calculated value then gets passed in to an invocation of the (compiled javascript) template function.

When compiling the compile-time value of a template argument, we look at its valueType and compile it as a string/runtime iterator/item property/etc accordingly.

We cannot yet pass in an item to a template, and then dereference an item property off of that item - it remains to be done.