Allow for composite values in XML properties. You can now have a element follow the mouse at a specified offset.
It's becoming clear that the parser is being too selective about what types of values can appear where. There should really just be statements and expressions. There are two type of statements: emit statements and handler statements. Expressions can be sub-components of statements, e.g. 4 + 5 is a composite statement of the two expressions 4 and 5, with the operator +.

The resolver should be the place where we enforce what type of value can appear in what place